Update index.html
I just created this branch.  I inserted a p tag about getting a 404 on this branch, it is not a real edit.
I'm doing this so I can make my first pull request.
My problem is when I try to see this webpage I'm getting a 404 and don't understand why.